:art: move fetch functions to db file
