+++ app/[usernameSlug]/evals/page.tsx
@@ -4,7 +4,7 @@ import { getProfileByUsername, getUser } from '@/db/profile'
 import { Project } from '@/db/project'
 import Link from 'next/link'
 import { Col } from '@/components/layout/col'
-import { ProfileTrust, ProjectEval } from '@/app/evals-form/page'
+import { ProfileTrust, ProjectEval } from '@/db/eval'
 import { ArrowLongRightIcon } from '@heroicons/react/20/solid'
 
 type Result = {
@@ -55,7 +55,7 @@ export default async function ResultsPage(props: {
   return (
     <div className="p-4">
       <h1 className="text-xl font-bold">{evaluator.full_name}&apos;s evals</h1>
-      <Col className="divide-y divide-gray-200">
+      <Col className="mt-5 divide-y divide-gray-200">
         <div className="grid grid-cols-7 gap-2 py-2 text-sm text-gray-600">
           <p className="col-span-3">Project</p>
           <p>Inside</p>
